backend/requirements.txt
django==3.1.7
djangorestframework==3.12.2
gunicorn==20.0.4
django-cors-headers==3.7.0
djangorestframework_simplejwt==4.6.0
# whitenoise==5.2.0
# django-impersonate==1.7.3
# django-crispy-forms==1.11.0
# crispy-bootstrap5==0.2

# django-auth-ldap==2.3.0
# django-spurl==0.6.7
# django-lifecycle==0.8.1

# pytest==6.2.2
# model-bakery==1.2.1
# pytest-cov==2.11.1
# pytest-mock==3.5.1
# pytest-django==4.1.0

# celery==5.0.5
# django-celery-beat==2.2.0
